Update Mobile Features
Mobile Features Page is in progress. The CSS is SASS compatible now. So far the header has been done. Will work on the making the rest a CSS Grid because that seems to be the best option.  Changed names of some CSS and added multiple files as well in order to update the features page :+1: